requirements.txt
fastapi==0.116.1
uvicorn==0.35.0

numpy==2.0.2
faiss-cpu==1.11.0.post1

sentence-transformers==5.1.0
transformers==4.55.4
torch==2.8.0

nltk==3.9.1

groq==0.31.0

pydantic==2.11.7